+++ src/main/resources/templates/student.html
@@ -5,18 +5,30 @@
     <title>Student</title>
 </head>
 <body>
-  <table>
+  <table border="1">
     <tr>
-      <td>编号</td><td th:text="studentInfo.serial"></td>
-      <td>姓名</td><td th:text="studentInfo.name"></td>
-      <td><a th:href="@{/class(className=${studentInfo.className})}">班级</a>></td><td th:text="studentInfo.className"></td>
-      <td>住址</td><td th:text="studentInfo.address"></td>
-      <td>年龄</td><td th:text="studentInfo.age"></td>
-      <td>语文</td><td th:text="studentInfo.language"></td>
-      <td>数学</td><td th:text="studentInfo.math"></td>
-      <td>英语</td><td th:text="studentInfo.english"></td>
-      <td>总成绩</td><td th:text="studentInfo.total"></td>
-      <td>平均成绩</td><td th:text="studentInfo.average"></td>
+      <td>编号</td>
+      <td>姓名</td>
+      <td><a th:href="@{/class(className=${studentInfo.className})}">班级</a></td>
+      <td>住址</td>
+      <td>年龄</td>
+      <td>语文</td>
+      <td>数学</td>
+      <td>英语</td>
+      <td>总成绩</td>
+      <td>平均成绩</td>
+    </tr>
+    <tr>
+      <td th:text="${studentInfo.getSerial()}"></td>
+      <td th:text="${studentInfo.getName()}"></td>
+      <td th:text="${studentInfo.getClassName()}"></td>
+      <td th:text="${studentInfo.getAddress()}"></td>
+      <td th:text="${studentInfo.getAge()}"></td>
+      <td th:text="${studentInfo.getLanguage()}"></td>
+      <td th:text="${studentInfo.getMaths()}"></td>
+      <td th:text="${studentInfo.getEnglish()}"></td>
+      <td th:text="${studentInfo.getTotal()}"></td>
+      <td th:text="${studentInfo.getAverage()}"></td>
     </tr>
   </table>
 </body>
