+++ src/main/resources/templates/studentQuery.html
@@ -5,7 +5,7 @@
     <title>Student Query</title>
 </head>
 <body>
-<form method="post" action="">
+<form method="post" action="/display/student">
   <table border="1">
     <tr>
       <td>Name: </td> <td><input name="studentName" id="studentName"></td>
